Hack a clear search button in a fun, but poorly composable way
Because the onClick function of `Clear` relies on the known behavior of
`handleInput`, if that function changes not only will this break, but it
constrains the reusability of this component outside this exact context. The
correct thing would have been ot have implenented a clear handling function that
relied on a name property that is shared between the input and the clear button.